Update whaspp.html
Hey Rishabh, that was a nice whatsapp design using HTML and CSS. However i found that some of the images are not loading the browser and it seems to be the error in giving the path in the <img src=""> tag. I have corrected all of them. I hope you find it useful.